Removed bot.py and moved former functionality to main.py
